Moved and documented plugin context provider and higher order component
