+++ edit-post/api/index.js
@@ -1,5 +1,5 @@
 import { PluginRegistry } from './plugin';
-import { PluginSidebarFill as PluginSidebar } from './components/plugin-sidebar';
+import { PluginSidebarFill as PluginSidebar } from '../components/plugin-sidebar';
 
 const registry = PluginRegistry.getInstance();
